feature: add mechanisms for event scheduling
Can put in a date and test to see if work unit report deadline is before that date.  Work unit priority sort includes priority favoring those below cutoff.  The date supplied at startup is re-calculated for confirmation an result is displayed as a message.  If date has error then no bunkering is enabled.